requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
dnspython==2.7.0
Flask==3.1.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pymongo==4.13.2
pypdf==5.7.0
python-dotenv==1.1.1
python-bidi==0.6.6
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
six==1.17.0
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
Flask-SocketIO==5.3.0
eventlet==0.40.4
weasyprint
